Move search results into new file. Misc tidying
